Top bar expanding now works with leap motion's mousemove
Only using mousemove to achieve the same effect of mouseleave. awwwww yeeeaaaah.
We did it by adding a mousemove listener to the body and check the event target, see if it is a child of the element, or the element itself.
If not, then it must be either parent element or another element, meaning the mouse is not on the element anymore